PES-76-shipping-methods (#9)

+++ install.zasilkovna.php
@@ -4,6 +4,10 @@
 defined('_JEXEC') or die('Restricted access');
 defined('DS') or define('DS', DIRECTORY_SEPARATOR);
 
+/**
+ * @param $src
+ * @param $dst
+ */
 function recurse_copy($src, $dst) {
 	if(is_dir($src)) {
 		$dir = opendir($src);
@@ -27,7 +31,11 @@ function recurse_copy($src, $dst) {
 	}
 }
 
-function recurse_delete($dir) {
+/**
+ * @param $dir
+ * @param false $ignore
+ */
+function recurse_delete($dir, $ignore = false) {
 	echo "deleting: " . $dir . "<br>";
 
 	if(is_dir($dir)) {
@@ -40,11 +48,21 @@ function recurse_delete($dir) {
 		rmdir($dir);
 	}
 	else {
-		unlink($dir);
+        if ($ignore) {
+            @unlink($dir);
+        } else {
+            unlink($dir);
+        }
 	}
 }
 
+/**
+ * Class plgVmShipmentZasilkovnaInstallerScript
+ */
 class plgVmShipmentZasilkovnaInstallerScript {
+
+    private $migratingPricingRules = false;
+
 	/**
 	 * Constructor
 	 *
@@ -66,6 +84,11 @@ class plgVmShipmentZasilkovnaInstallerScript {
 
 	}
 
+    /**
+     * @param $haystack
+     * @param $needle
+     * @return bool
+     */
     function endsWith($haystack, $needle)
     {
         $length = strlen($needle);
@@ -86,8 +109,8 @@ class plgVmShipmentZasilkovnaInstallerScript {
 	 */
 	public function postflight($route, JAdapterInstance $adapter) {
 		$vm_admin_path = JPATH_ADMINISTRATOR . DS . 'components' . DS . 'com_virtuemart';
+        $media_path = JPATH_ROOT . DS . 'media' . DS . 'com_zasilkovna' . DS;
 		if($route == "install") {
-			$media_path = JPATH_ROOT . DS . 'media' . DS . 'com_zasilkovna' . DS;
 			recurse_copy($media_path . 'admin' . DS . 'com_virtuemart' . DS, $vm_admin_path . DS);
 
             $files = scandir($media_path . 'admin' . DS);
@@ -195,7 +218,188 @@ INSERT INTO #__virtuemart_adminmenuentries (`module_id`, `parent_id`, `name`, `l
             }
         }
 
+        $this->migratePricingRules();
 	}
+
+    /**
+     *  migrates price rules
+     */
+    private function migratePricingRules() {
+        require_once JPATH_ADMINISTRATOR . '/components/com_virtuemart/install/script.virtuemart.php';
+        $vmInstall = new \com_virtuemartInstallerScript();
+        $vmInstall->loadVm(false);
+
+        require_once __DIR__ . '/zasilkovna.php';
+
+        /** @var \VirtueMartModelZasilkovna $model */
+        $model = VmModel::getModel('zasilkovna');
+        $config = $model->loadConfig();
+        if ($this->migratingPricingRules || empty($config) || isset($config['pricing_rules_migration_completed_at'])) {
+            return;
+        }
+
+        $this->migratingPricingRules = true;
+        echo 'Migrating pricing rules.<br>';
+
+        /** @var \VirtueMartModelShipmentmethod $shipmentmethodModel */
+        $shipmentmethodModel = VmModel::getModel('shipmentmethod');
+
+        $globalWeightRules = [];
+        $globalMaxWeight = $model->getConfig('global/values/maximum_weight', '');
+        $globalShipmentCost = $model->getConfig('global/values/default_price', '');
+        $globalFreeShipment = $model->getConfig('global/values/free_shipping', '');
+
+        if (!is_numeric($globalShipmentCost)) {
+            echo 'Global configuration not found. Migration stopped.<br>';
+
+            $config = $model->loadConfig();
+            $config['pricing_rules_migration_completed_at'] = (new \DateTime())->format(\DateTime::ISO8601);
+            $model->updateConfig($config);
+            return;
+        }
+
+        $pricingRules = [];
+        $countries = ['cz', 'sk', 'hu', 'ro', 'pl'];
+        $countriesWithOther = array_merge(['other'], $countries); // is used to search in config
+        $countryIds = [];
+
+        foreach ($countries as $country) {
+            /** @var VmTable $countryObject */
+            $countryObject = \VirtueMartModelCountry::getCountry(strtoupper($country), 'country_2_code');
+            $countryIds[$country] = $countryObject->virtuemart_country_id;
+        }
+
+        $pricingRulesCount = 0;
+        foreach ($countriesWithOther as $country) {
+            $countryDefaultPrice = $model->getConfig($country. '/values/default_price', '');
+            $countryFreeShipment = $model->getConfig($country. '/values/free_shipping', '');
+
+            $countryWeightRules = ($model->getConfig($country) ?: []);
+            unset($countryWeightRules['values']);
+
+            usort($countryWeightRules, function ($countryWeightRulesA, $countryWeightRuleB) {
+                if ($countryWeightRulesA['weight_from'] === $countryWeightRuleB['weight_from']) {
+                    return 0;
+                }
+
+                return ($countryWeightRulesA['weight_from'] > $countryWeightRuleB['weight_from'] ? 1 : -1);
+            });
+
+            $lastCountryWeightRule = null;
+            $countryWeightRulesTransformed = [];
+            $weightRulesCount = 0;
+            foreach ($countryWeightRules as $countryWeightRule) {
+
+                $addWeightRule = function (&$countryWeightRulesTransformed) use ($countryWeightRule, $lastCountryWeightRule, &$weightRulesCount, $countryDefaultPrice, $globalShipmentCost) {
+                    if ($lastCountryWeightRule === null && $countryWeightRule['weight_from'] > 0) {
+                        $key = 'weightRules' . $weightRulesCount;
+                        $countryWeightRulesTransformed[$key] = [
+                            'maxWeightKg' => $countryWeightRule['weight_from'],
+                            'price' => ($countryDefaultPrice ?: $globalShipmentCost),
+                        ];
+                        $weightRulesCount++;
+                    }
+
+                    if (!empty($countryWeightRule['weight_from']) && $lastCountryWeightRule && $lastCountryWeightRule['weight_to'] != $countryWeightRule['weight_from']) {
+                        $key = 'weightRules' . $weightRulesCount;
+                        $countryWeightRulesTransformed[$key] = [
+                            'maxWeightKg' => $countryWeightRule['weight_from'],
+                            'price' => ($countryDefaultPrice ?: $globalShipmentCost),
+                        ];
+                        $weightRulesCount++;
+                    }
+
+                    if (!empty($countryWeightRule['weight_to'])) {
+                        $key = 'weightRules' . $weightRulesCount;
+                        $countryWeightRulesTransformed[$key] = [
+                            'maxWeightKg' => $countryWeightRule['weight_to'],
+                            'price' => (($countryWeightRule['price'] ?: $countryDefaultPrice) ?: $globalShipmentCost),
+                        ];
+                        $weightRulesCount++;
+                    }
+                };
+
+                if ($country === 'other') {
+                    $addWeightRule($globalWeightRules);
+                } else {
+                    $addWeightRule($countryWeightRulesTransformed);
+                }
+
+                $lastCountryWeightRule = $countryWeightRule;
+            }
+
+            if (!is_numeric($countryDefaultPrice) && !is_numeric($countryFreeShipment) && empty($countryWeightRulesTransformed)) {
+                continue; // no usable data to migrate
+            }
+
+            if ($country === 'other') {
+                continue; // other country default price and free shipping will be lost
+            }
+
+            $pricingRules['pricingRules' . $pricingRulesCount] = [
+                'country' => $countryIds[$country],
+                'shipment_cost' => $countryDefaultPrice,
+                'free_shipment' => $countryFreeShipment,
+                'weightRules' => $countryWeightRulesTransformed
+            ];
+
+            $pricingRulesCount++;
+        }
+
+        $shipmentMethodIds = $model->getShipmentMethodIds();
+        $model->publishShipmentMethods($shipmentMethodIds, 0);
+        echo 'All packetery methods were unpublished. Please check them and publish them again.<br>';
+
+        $packetaMethod = \VirtueMartModelZasilkovna\ShipmentMethod::fromRandom(
+            [
+                'globalWeightRules' => $globalWeightRules,
+                'pricingRules' => $pricingRules
+            ]
+        );
+        $sortedMethod = $packetaMethod->getResortedClone()->toArray();
+        $globalWeightRules = $sortedMethod['globalWeightRules'];
+        $pricingRules = $sortedMethod['pricingRules'];
+
+        $pricingRulesEncoded = json_encode($pricingRules);
+        $globalWeightRulesEncoded = json_encode($globalWeightRules);
+        $params = [
+            "maxWeight=\"$globalMaxWeight\"",
+            "shipment_cost=\"$globalShipmentCost\"",
+            "free_shipment=\"$globalFreeShipment\"",
+            "globalWeightRules=$globalWeightRulesEncoded",
+            "pricingRules=$pricingRulesEncoded"
+        ];
+
+        $data = [
+            'shipment_name' => 'Packeta PP',
+            'shipment_desc' => '',
+            'shipment_element' => VirtueMartModelZasilkovna::PLG_NAME,
+            'shipment_jplugin_id' => $model->getExtensionId(),
+            'shipment_params' => implode('|', $params),
+            'published' => 0,
+        ];
+
+        $table = $shipmentmethodModel->getTable('shipmentmethods');
+        $table->bindChecknStore($data);
+
+        $xrefTable = $shipmentmethodModel->getTable('shipmentmethod_shoppergroups');
+        $xrefTable->bindChecknStore($data);
+
+        $id = $data['virtuemart_shipmentmethod_id'];
+
+        $db =& JFactory::getDBO();
+        $q = "UPDATE #__virtuemart_shipmentmethods SET shipment_params='" . $data['shipment_params'] . "' WHERE virtuemart_shipmentmethod_id='{$id}'";
+        $db->setQuery($q);
+        $db->execute();
+
+        echo 'New shipment method with migrated rules was created.<br>';
+
+        $config = $model->loadConfig();
+        $config['pricing_rules_migration_completed_at'] = (new \DateTime())->format(\DateTime::ISO8601);
+        $model->updateConfig($config);
+
+        $this->migratingPricingRules = false;
+    }
 
 	/**
 	 * Called on uninstallation
@@ -211,14 +415,15 @@ INSERT INTO #__virtuemart_adminmenuentries (`module_id`, `parent_id`, `name`, `l
 		$vm_admin_path = JPATH_ADMINISTRATOR . DS . 'components' . DS . 'com_virtuemart';
 		recurse_delete($vm_admin_path . DS . 'models' . DS . 'zasilkovna.php');
 		recurse_delete($vm_admin_path . DS . 'models' . DS . 'zasilkovna_orders.php');
+		recurse_delete($vm_admin_path . DS . 'models' . DS . 'zasilkovna_src' . DS);
 		recurse_delete($vm_admin_path . DS . 'views' . DS . 'zasilkovna' . DS);
 		recurse_delete($vm_admin_path . DS . 'controllers' . DS . 'zasilkovna.php');
 		recurse_delete(JPATH_ADMINISTRATOR . DS . 'language' . DS . 'en-GB' . DS . 'en-GB.plg_vmshipment_zasilkovna.ini');
 		recurse_delete(JPATH_ADMINISTRATOR . DS . 'language' . DS . 'cs-CZ' . DS . 'cs-CZ.plg_vmshipment_zasilkovna.ini');
-		recurse_delete(JPATH_ADMINISTRATOR . DS . 'language' . DS . 'sk-SK' . DS . 'sk-SK.plg_vmshipment_zasilkovna.ini');
-		recurse_delete(JPATH_ADMINISTRATOR . DS . 'language' . DS . 'pl-PL' . DS . 'pl-PL.plg_vmshipment_zasilkovna.ini');
-		recurse_delete(JPATH_ADMINISTRATOR . DS . 'language' . DS . 'hu-HU' . DS . 'hu-HU.plg_vmshipment_zasilkovna.ini');
-		recurse_delete(JPATH_ADMINISTRATOR . DS . 'language' . DS . 'ro-RO' . DS . 'ro-RO.plg_vmshipment_zasilkovna.ini');                                
+		recurse_delete(JPATH_ADMINISTRATOR . DS . 'language' . DS . 'sk-SK' . DS . 'sk-SK.plg_vmshipment_zasilkovna.ini', true);
+		recurse_delete(JPATH_ADMINISTRATOR . DS . 'language' . DS . 'pl-PL' . DS . 'pl-PL.plg_vmshipment_zasilkovna.ini', true);
+		recurse_delete(JPATH_ADMINISTRATOR . DS . 'language' . DS . 'hu-HU' . DS . 'hu-HU.plg_vmshipment_zasilkovna.ini', true);
+		recurse_delete(JPATH_ADMINISTRATOR . DS . 'language' . DS . 'ro-RO' . DS . 'ro-RO.plg_vmshipment_zasilkovna.ini', true);
 	}
 
 }
